Add missing 'toJSON' implementations.
I didn't realize that 'toJSON' and 'toEncoding' were separate functions.
The 'toEncoding' implementation for the 'Notification' type was properly
converting _-style accessor names, but as 'toJSON' was using the default
implementation, the latter was not. Therefore, the JSON values produced
by 'toEncoding' and 'toJSON' were different, and this was the source of
the failed 'validateToJSON' Swagger schema test.