Make TopicName and TopicId proper newtypes.
Also switch their meaning around; the app now returns TopicId's upon
creation of a Topic. TopicName is the service-specific name.

Note that, currently, generation of the TopicId from a Topic is stupid
(it's just the text of the TopicName) and will break as soon as we add
support for another service. A proper implementation should hash the
topic name and service name somehow.