Back to basics.
Originally, this project was supposed to be just a simple doorbell
service that used Amazon SNS to notify subscribers that the bell had
been rung. It started evolving into something more -- basically, a proxy
service for creating topics, sending notifications, and handling
multiple pubsub/messaging services.

I've decided I don't want to deal with all of that at this time, and
just want to get a doorbell service running ASAP. This commit returns
the project to the original scope.

Note that the topic ARN is currently hard-wired and will need to be made
configurable.