Preliminary support for APNS messages via SNS.
The interfaces need work, but pretty much everything here is
working... except for the fact that, when delivering JSON to APNS via
an Amazon SNS message, SNS requires the APNS JSON to be sent as a raw
string, not as a nested JSON object. In other words, rather than
sending this string to SNS:

{
  "default" : "Bar",
  "APNS_SANDBOX" :
    {
      "aps" :
        {
          "alert" :
            {
              "title" : "Foo",
              "body" :"Ring! Ring!"
            },
          "sound" : "default"
        }
    }
}

we actually need to send this string to SNS:

{
  "default" : "Bar",
  "APNS_SANDBOX" : "{\"aps\":{\"alert\":{\"title\":\"Foo\",\"body\":\"Ring! Ring!\"},\"sound\":\"default\"}}"
}

Gross.

Currently, the wire formatter does the former, not the latter. This
will be fixed in a subsequent commit.