Proper APNS encoding.
Fixes the issue mentioned in the previous commit (APNS JSON is now
encoded as a string, rather than as a nested object).

Note that the Apple APNS docs state that JSON \U notation is not
supported, so I need to test UTF-8 character encoding as I'm pretty
certain that Aeson will generate \U notation per the standard.