Make sure haskellPackages.pinpon doesn't run hlint tests.
When upstream packages use this overlay, we want to make sure that
hlint issues don't cause build failures, but at the same time, we want
to run hlint tests on our own CI builds for this repo. Therefore, we
set haskellPackages.pinpon to be the non-hlint version, and add a new
package, haskellPackages.pinponHlint, which is what the jobsets for
this repo will build.